moved parser constructor traits (parsec/Parsec*.h) into parsec/parser/
