+++ include/cparsec3/parsec/parser/ParsecPrim1.h
@@ -1,7 +1,9 @@
 /* -*- coding: utf-8-unix -*- */
-#include "../base/base_generics.h"
+#pragma once
 
-#include "parsec.h"
+#include "../../base/base_generics.h"
+
+#include "../parsec.h"
 
 // -----------------------------------------------------------------------
 #define ParsecPrim1(...) TYPE_NAME(ParsecPrim, __VA_ARGS__)
